Added temperature calibration and additional confirmation dialog for factory reset.

+++ code/main.h
@@ -66,7 +66,7 @@ struct calibration {
 	uint16_t zeroC;
 	uint16_t hundredC;
 	uint16_t offset;
-	uint16_t coefficient;
+	float    coefficient;
 } *stCalibration;
 
 struct periods {
